requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
crispy-bootstrap5==0.5
dj-database-url==0.5.0
Django==3.2.5
django-crispy-forms==1.12.0
django-embed-video==1.4.0
djangorestframework==3.11.0
gunicorn==20.1.0
idna==3.2
Pillow==8.3.2
psycopg2==2.9.1
pytz==2021.1
requests==2.26.0
rest-framework-generic-relations==2.0.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.3.0
